backend-agentic/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "WebSocket streaming backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=12",
    "langgraph>=0.2",
    "pydantic>=2",
    "requests>=2",
    "python-dotenv>=1.0",
    "psutil>=5.9"
]

[dependency-groups]
dev = []
